refactor helpers out of processor.base
