+++ ocrd/ocrd/processor/__init__.py
@@ -1,5 +1,7 @@
 from .base import (
     Processor,
+)
+from .helpers import (
     run_cli,
     run_processor,
     generate_processor_help
